2309 to Final (#2867)
Review period has ended. Moving this EIP to Final! This EIP was originally proposed in October of 2019. It has been refined and tested in production environments. It recently helped facilitate the creation of 20,000 NFT vouchers where 500 of these vouchers were sold in a pre-sale. By including the event mentioned in this EIP we are able to effectively track the creation and distribution of these vouchers. Vouchers could be purchased in bulk and tracked effectively - saving consumers massive gas costs for bulk purchases.